backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pdfplumber==0.10.2
python-docx==0.8.11
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
